Add drawKarnaughMap
