+++ src/js/truthtrick.mjs
@@ -6,78 +6,12 @@ import getOutputs from './outputs';
 
 import { drawCircuit, resize } from './drawCircuit';
 import drawTruthTable from './drawTruthTable';
+import drawKarnaughMap from './drawKarnaughMap';
 
 // The input element for the expression
 const expression = document.getElementById('expressionInput');
-// The table element for the karnaugh map
-const karnaughMap = document.getElementById('karnaugh-map');
-// The section element for the karnaugh map
-const karnaughMapSection = document.getElementById('karnaugh-map-section');
 // The list of operator buttons
 const operationButtons = document.getElementsByClassName('operation');
-
-// Returns a value n_i that differs by one bit from n_(i-1)
-// gray(0) = 0, gray(1) = 1, gray(2) = 3
-const gray = i => i ^ (i >> 1); // eslint-disable-line no-bitwise
-
-// Creates a karnaugh map
-const createKarnaughMap = (outputs, variables) => {
-  const vertical = Math.floor(variables.length / 2);
-  const horizontal = variables.length - vertical;
-
-  // Hide karnaugh map section if there is only one variable
-  if (variables.length <= 1) {
-    karnaughMapSection.hidden = true;
-    return;
-  }
-
-  karnaughMapSection.hidden = false;
-
-  karnaughMap.firstElementChild.innerHTML = '';
-
-  const header = karnaughMap.insertRow();
-
-  // Creates top left corner
-  const key = header.insertCell();
-  const keyDiv = key.appendChild(document.createElement('div'));
-  keyDiv.classList.add('overflow');
-  const keyText = `${variables.slice(0, vertical).join('')}\\${variables.slice(vertical).join('')}`;
-  keyDiv.appendChild(document.createTextNode(keyText));
-
-  // Creates headings for the top
-  for (let i = 0; i < 2 ** horizontal; i += 1) {
-    const cell = header.insertCell();
-    const div = cell.appendChild(document.createElement('div'));
-    div.classList.add('overflow');
-    div.appendChild(document.createTextNode(gray(i).toString(2).padStart(horizontal, '0')));
-  }
-
-  // Creates headings for the left
-  for (let i = 0; i < 2 ** vertical; i += 1) {
-    const row = karnaughMap.insertRow();
-    const firstCell = row.insertCell();
-    const firstDiv = firstCell.appendChild(document.createElement('div'));
-    firstDiv.classList.add('overflow');
-    firstDiv.appendChild(document.createTextNode(gray(i).toString(2).padStart(vertical, '0')));
-
-    // Fills in the rest of the cells
-    for (let j = 0; j < 2 ** horizontal; j += 1) {
-      const index = gray(i).toString(2).padStart(vertical, '0') + gray(j).toString(2).padStart(horizontal, '0');
-      const digit = outputs[parseInt(index, 2)];
-      const cell = row.insertCell();
-      const div = cell.appendChild(document.createElement('div'));
-      div.classList.add('overflow');
-      div.appendChild(document.createTextNode(digit));
-
-      // Highlights 1s
-      if (digit === '1') {
-        cell.classList.add('highlighted');
-      } else {
-        cell.classList.remove('highlighted');
-      }
-    }
-  }
-};
 
 // This runs whenever the expression is changed
 const newExpression = () => {
@@ -89,7 +23,7 @@ const newExpression = () => {
 
   drawCircuit(exp);
   drawTruthTable(inputs, outputs, variables);
-  createKarnaughMap(outputs, variables);
+  drawKarnaughMap(outputs, variables);
 };
 
 // This runs whenever an operator button is clicked
